+++ test/typed_json_test.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:test/test.dart';
+import 'package:serializer/codecs.dart';
 import 'package:serializer/serializer.dart';
 
 
@@ -85,9 +86,14 @@ class TypedWithIgnore extends TypedProxyA {
   TypedWithIgnore([this.a, this.b, this.secret]);
 }
 
+Serializer _dateSerializer = new Serializer.TypedJson()
+    ..addTypeCodec(DateTime, new DateTimeCodec());
 @serializable
 class TypedDate extends TypedProxyA {
   DateTime date = new DateTime.now();
+
+  @ignore
+  Serializer get serializer => _dateSerializer;
 
   TypedDate([this.date]);
 }
@@ -152,7 +158,8 @@ class Address {
 }
 
 main() {
-  var serializer = new Serializer.TypedJson();
+  var serializer = new Serializer.TypedJson()
+      ..addTypeCodec(DateTime, new DateTimeCodec());
 
 
   group("Serialize", () {